Refactor params
